[update] Refactoring to App.js
